requirements.txt
scipy
numpy
tensorflow-gpu>=1.5
